+++ tests/test_edge_cases_dag.py
@@ -2,13 +2,27 @@ from functools import partial
 from typing import Any, Tuple
 
 import pytest
+from networkx import NetworkXUnfeasible
 from tawazi import DAG, dag, xn
 from tawazi._dag.digraph import DiGraphEx
 from tawazi._dag.helpers import execute
+from tawazi.node import ExecNode, UsageExecNode
 
 
 def shortcut_execute(dag: DAG[Any, Any], graph: DiGraphEx) -> Any:
     return execute(exec_nodes=dag.exec_nodes, max_concurrency=dag.max_concurrency, graph=graph)
+
+
+def a(c: str) -> str:
+    return "a"
+
+
+def b(a: str) -> str:
+    return a + "b"
+
+
+def c(b: str) -> str:
+    return b + "c"
 
 
 def func(a: str, b: str) -> str:
@@ -63,3 +77,17 @@ def test_setup_debug_nodes() -> None:
         @xn(debug=True, setup=True)
         def a() -> None:
             ...
+
+
+def test_circular_deps() -> None:
+    en_a = ExecNode(id_="a", exec_function=a, args=[], is_sequential=True)
+    en_b = ExecNode(
+        id_="b", exec_function=b, args=[UsageExecNode(en_a.id)], priority=2, is_sequential=False
+    )
+    en_c = ExecNode(
+        id_="c", exec_function=c, args=[UsageExecNode(en_a.id)], priority=1, is_sequential=False
+    )
+    en_a.args = [UsageExecNode(en_c.id)]
+
+    with pytest.raises(NetworkXUnfeasible):
+        DAG({xn.id: xn for xn in [en_a, en_b, en_c]}, [], [], 2)
